pyproject.toml
[tool.poetry]
name = "dashboard_logistica"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.31.1"
plotly = "^5.19.0"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.2"
streamlit-autorefresh = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
